requirements.txt
PyYAML==3.11
beautifulsoup4==4.3.2
lxml==3.3.5
pathtools==0.1.2
watchdog==0.8.0
